Resolve #4 - add MTOM support
JAXWSEnvironment now supports MTOM. Flag for enabling MTOM was added to
AbstractBuilder. In order to enable MTOM support for clients, enableMtom
method has to be invoked on ClientBuilder. To enable MTOM support for
server endpoints, invoke enableMtom on EndpointBuilder. If you are using
@MTOM JAX-WS annotation on service implementation, enableMtom does not
have to be invoked on EndpointBuilder.

MTOM example was added to JaxWsExampleApplication. MTOM example is a WSDL
first service (MtomService.wsdl and MtomServiceImpl.java). Client example
is implemented in RESTful resource AccessMtomServiceResource.java.